Fix circular reference in http
